requirements.txt
reflex==0.3.2
gspread==5.12.0
google-auth==2.23.4
requests==2.32.4
notion-client==2.7.0
google-generativeai==0.3.2
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0
playwright-stealth==2.0.1
python-dotenv==1.0.0
tqdm>=4.66.0